Move `zbstorehttp.Store.PutObject` logic to separate file
It's about to get more complex.

Updates #260
